PR Comments

+++ contracts/tests/ServiceRegistryV2.sol
@@ -216,7 +216,7 @@ contract ServiceRegistryV2 is Initializable, ERC2771RecipientUpgradeable, UUPSUp
      * @param _serviceDataUri token Id to IPFS URI mapping
      */
     function createOpenServiceFromBuyer(uint256 _platformId, string calldata _serviceDataUri) public payable returns (uint256) {
-        uint16 servicePostingFee = talentLayerPlatformIdContract.getServicePostingFee(_platformId);
+        uint256 servicePostingFee = talentLayerPlatformIdContract.getServicePostingFee(_platformId);
         require(msg.value == servicePostingFee, "Non-matching funds");
         talentLayerPlatformIdContract.isValid(_platformId);
         uint256 senderId = tlId.walletOfOwner(_msgSender());
@@ -240,7 +240,7 @@ contract ServiceRegistryV2 is Initializable, ERC2771RecipientUpgradeable, UUPSUp
         uint256 senderId = tlId.walletOfOwner(_msgSender());
         require(senderId > 0, "You should have a TalentLayerId");
         require(allowedTokenList[_rateToken], "This token is not allowed");
-        uint16 proposalPostingFee = talentLayerPlatformIdContract.getProposalPostingFee(_platformId);
+        uint256 proposalPostingFee = talentLayerPlatformIdContract.getProposalPostingFee(_platformId);
         require(msg.value == proposalPostingFee, "Non-matching funds");
 
         Service storage service = services[_serviceId];
