+++ contracts/tests/TalentLayerPlatformIDV2.sol
@@ -35,8 +35,8 @@ contract TalentLayerPlatformIDV2 is ERC721Upgradeable, AccessControlUpgradeable,
         string dataUri;
         uint16 originServiceFeeRate;
         uint16 originValidatedProposalFeeRate;
-        uint16 servicePostingFee;
-        uint16 proposalPostingFee;
+        uint256 servicePostingFee;
+        uint256 proposalPostingFee;
         Arbitrator arbitrator;
         bytes arbitratorExtraData;
         uint256 arbitrationFeeTimeout;
@@ -134,7 +134,7 @@ contract TalentLayerPlatformIDV2 is ERC721Upgradeable, AccessControlUpgradeable,
      * @param _platformId Platform Id to check
      * @return The Service posting fee
      */
-    function getServicePostingFee(uint256 _platformId) external view returns (uint16) {
+    function getServicePostingFee(uint256 _platformId) external view returns (uint256) {
         require(_platformId > 0 && _platformId < _nextTokenId.current(), "Invalid platform ID");
         return platforms[_platformId].servicePostingFee;
     }
@@ -144,7 +144,7 @@ contract TalentLayerPlatformIDV2 is ERC721Upgradeable, AccessControlUpgradeable,
      * @param _platformId Platform Id to check
      * @return The Proposal posting fee
      */
-    function getProposalPostingFee(uint256 _platformId) external view returns (uint16) {
+    function getProposalPostingFee(uint256 _platformId) external view returns (uint256) {
         require(_platformId > 0 && _platformId < _nextTokenId.current(), "Invalid platform ID");
         return platforms[_platformId].proposalPostingFee;
     }
@@ -287,7 +287,7 @@ contract TalentLayerPlatformIDV2 is ERC721Upgradeable, AccessControlUpgradeable,
      * @param _platformId The platform id
      * @param _servicePostingFee The new fee
      */
-    function updateServicePostingFee(uint256 _platformId, uint16 _servicePostingFee) public {
+    function updateServicePostingFee(uint256 _platformId, uint256 _servicePostingFee) public {
         require(ownerOf(_platformId) == msg.sender, "You're not the owner of this platform");
 
         platforms[_platformId].servicePostingFee = _servicePostingFee;
@@ -299,7 +299,7 @@ contract TalentLayerPlatformIDV2 is ERC721Upgradeable, AccessControlUpgradeable,
      * @param _platformId The platform id
      * @param _proposalPostingFee The new fee
      */
-    function updateProposalPostingFee(uint256 _platformId, uint16 _proposalPostingFee) public {
+    function updateProposalPostingFee(uint256 _platformId, uint256 _proposalPostingFee) public {
         require(ownerOf(_platformId) == msg.sender, "You're not the owner of this platform");
 
         platforms[_platformId].proposalPostingFee = _proposalPostingFee;
@@ -548,11 +548,11 @@ contract TalentLayerPlatformIDV2 is ERC721Upgradeable, AccessControlUpgradeable,
      * @notice Emit when the service posting fee is updated for a platform
      * @param _servicePostingFee The new fee
      */
-    event ServicePostingFeeUpdated(uint256 _platformId, uint16 _servicePostingFee);
+    event ServicePostingFeeUpdated(uint256 _platformId, uint256 _servicePostingFee);
 
     /**
      * @notice Emit when the proposal posting fee is updated for a platform
      * @param _proposalPostingFee The new fee
      */
-    event ProposalPostingFeeUpdated(uint256 _platformId, uint16 _proposalPostingFee);
+    event ProposalPostingFeeUpdated(uint256 _platformId, uint256 _proposalPostingFee);
 }
